rename front-matter to markdown
