+++ packages/embodi/src/core/dev.ts
@@ -1,6 +1,6 @@
 import { createServer, preview, defineConfig, type Plugin } from 'vite';
 import { configPlugin, devServerPlugin, virtualPlugin } from './vite/embodi.js';
-import { embodiFrontMatter } from './vite/front-matter.js';
+import { embodiFrontMatter } from './vite/markdown.js';
 import { svelte, vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 import { embodiSvelte } from './vite/svelte.js';
